backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
httpx==0.27.0
python-dateutil==2.9.0
reportlab==4.1.0
Pillow==10.3.0
langdetect==1.0.9
anthropic>=0.86.0
torch>=2.0.0
transformers>=4.30.0
pytesseract>=0.3.10
pytest==8.2.0
pytest-asyncio==0.23.6
